io.py -> api.py: Refactored into API for JS frontend
RESTful API
-----------

A JSON list of object keys is obtained via

    /api/Documents
    /api/Documents/doc_id/Threads
    /api/Documents/doc_id/Threads/thread_id/Revisions
    /api/Users

Posting an object to these collections wipes its original
ID and replaces it with a newly generated one, creating
a new and identical copy.

A JSON dictionary of object properties is obtained via

    /api/Documents/doc_id
    /api/Documents/doc_id/Threads/thread_id
    /api/Documents/doc_id/Threads/thread_id/Revisions/rev_id

A put command creates/replaces an object in the set.

TODO
----

* Write test cases
* Voting is not working yet
* Neither is forking